addCreator {openxlsx}	R Documentation

Add another author to the meta data of the file.

Description

Just a wrapper of wb$addCreator()

Usage

addCreator(wb, Creator)

Arguments

wb
A workbook object

Creator
A string object with the name of the creator

Author(s)

Philipp Schauberger

Examples

wb <- createWorkbook()
addCreator(wb, "test")

[Package openxlsx version 4.2.5 Index]