Add column filters

Add excel column filters to a worksheet

addFilter(wb, sheet, rows, cols)

Arguments

wb
A workbook object

sheet
A name or index of a worksheet

rows
A row number.

cols
columns to add filter to.

Details

adds filters to worksheet columns, same as filter parameters in writeData. writeDataTable automatically adds filters to first row of a table. NOTE Can only have a single filter per worksheet unless using tables.

See also

Examples

wb <- createWorkbook()addWorksheet(wb, "Sheet 1")addWorksheet(wb, "Sheet 2")addWorksheet(wb, "Sheet 3")writeData(wb, 1, iris)addFilter(wb, 1, row = 1, cols = 1:ncol(iris))writeData(wb, 2, x = iris, withFilter = TRUE)writeDataTable(wb, 3, iris)if (FALSE) saveWorkbook(wb, file = "addFilterExample.xlsx", overwrite = TRUE)