conditionalFormat {openxlsx}	R Documentation

Add conditional formatting to cells

Description

DEPRECATED! USE conditionalFormatting()

Usage

conditionalFormat(
  wb,
  sheet,
  cols,
  rows,
  rule = NULL,
  style = NULL,
  type = "expression"
)

Arguments

wb
A workbook object

sheet
A name or index of a worksheet

cols
Columns to apply conditional formatting to

rows
Rows to apply conditional formatting to

rule
The condition under which to apply the formatting or a vector of colours. See examples.

style
A style to apply to those cells that satisfy the rule. A Style object returned from createStyle()

type
Either 'expression', 'colorscale' or 'databar'. If 'expression' the formatting is determined by a formula. If colorScale cells are coloured based on cell value. See examples.

Details

DEPRECATED! USE conditionalFormatting()

Valid operators are "<", "<=", ">", ">=", "==", "!=". See Examples. Default style given by: createStyle(fontColour = "#9C0006", bgFill = "#FFC7CE")

Author(s)

Alexander Walker

See Also

createStyle()

[Package openxlsx version 4.2.5 Index]