convertToDate {openxlsx}	R Documentation

Convert from excel date number to R Date type

Description

Convert from excel date number to R Date type

Usage

convertToDate(x, origin = "1900-01-01", ...)

Arguments

x
A vector of integers

origin
date. Default value is for Windows Excel 2010

...
additional parameters passed to as.Date()

Details

Excel stores dates as number of days from some origin day

See Also

writeData()

Examples

## 2014 April 21st to 25th
convertToDate(c(41750, 41751, 41752, 41753, 41754, NA))
convertToDate(c(41750.2, 41751.99, NA, 41753))

[Package openxlsx version 4.2.5 Index]