createComment {openxlsx}	R Documentation

create a Comment object

Description

Create a cell Comment object to pass to writeComment()

Usage

createComment(
  comment,
  author = Sys.getenv("USERNAME"),
  style = NULL,
  visible = TRUE,
  width = 2,
  height = 4
)

Arguments

comment
Comment text. Character vector.

author
Author of comment. Character vector of length 1

style
A Style object or list of style objects the same length as comment vector. See createStyle().

visible
TRUE or FALSE. Is comment visible.

width, height
Width and height of textbook (in number of cells); doubles are rounded with base::round()

See Also

writeComment()

Examples

wb <- createWorkbook()
addWorksheet(wb, "Sheet 1")

c1 <- createComment(comment = "this is comment")
writeComment(wb, 1, col = "B", row = 10, comment = c1)

s1 <- createStyle(fontSize = 12, fontColour = "red", textDecoration = c("BOLD"))
s2 <- createStyle(fontSize = 9, fontColour = "black")

c2 <- createComment(comment = c("This Part Bold red\n\n", "This part black"), style = c(s1, s2))
c2

writeComment(wb, 1, col = 6, row = 3, comment = c2)
## Not run: 
saveWorkbook(wb, file = "createCommentExample.xlsx", overwrite = TRUE)

## End(Not run)

[Package openxlsx version 4.2.5 Index]