makeHyperlinkString {openxlsx}	R Documentation

create Excel hyperlink string

Description

Wrapper to create internal hyperlink string to pass to writeFormula()

Usage

makeHyperlinkString(sheet, row = 1, col = 1, text = NULL, file = NULL)

Arguments

sheet
Name of a worksheet

row
integer row number for hyperlink to link to

col
column number of letter for hyperlink to link to

text
display text

file
Excel file name to point to. If NULL hyperlink is internal.

See Also

writeFormula

[Package openxlsx version 4.1.4 Index]