setWindowSize {openxlsx}	R Documentation

Set and Get Window Size for xlsx file

Description

Set and Get Window Size for xlsx file

Usage

setWindowSize(
  wb,
  xWindow = NULL,
  yWindow = NULL,
  windowWidth = NULL,
  windowHeight = NULL
)

getWindowSize(wb)

Arguments

wb
A Workbook object

xWindow
the horizontal coordinate of the top left corner of the window

yWindow
the vertical coordinate of the top left corner of the window

windowWidth
the width of the window

windowHeight
the height of the window

Set the size and position of the window when you open the xlsx file. The units are in twips. See Microsoft's documentation for the xlsx standard

Examples

## Create Workbook object and add worksheets
wb <- createWorkbook()
addWorksheet(wb, "S1")
getWindowSize(wb)
setWindowSize(wb, windowWidth = 10000)

[Package openxlsx version 4.2.8.1 Index]