create a Comment object

Create a cell Comment object to pass to writeComment()

createComment(
  comment,
  author = Sys.getenv("USERNAME"),
  style = NULL,
  visible = TRUE,
  width = 2,
  height = 4
)

Arguments

comment
Comment text. Character vector.

author
Author of comment. Character vector of length 1

style
A Style object or list of style objects the same length as comment vector. See createStyle.

visible
TRUE or FALSE. Is comment visible.

width
Textbox integer width in number of cells

height
Textbox integer height in number of cells

See also

Examples

wb <- createWorkbook()
addWorksheet(wb, "Sheet 1")

c1 <- createComment(comment = "this is comment")
writeComment(wb, 1, col = "B", row = 10, comment = c1)

s1 <- createStyle(fontSize = 12, fontColour = "red", textDecoration = c("BOLD"))
s2 <- createStyle(fontSize = 9, fontColour = "black")

c2 <- createComment(comment = c("This Part Bold red\n\n", "This part black"), style = c(s1, s2))
c2

#> Author: 
#> Text:
#>  This Part Bold red
#> 
#> This part black
#> 
#> Style:
#> Font size: 12
#> Font colour: #FF0000
#> Font decoration: BOLD
#> 
#> 
#> Font size: 9
#> Font colour: #000000
#> 
#> 

writeComment(wb, 1, col = 6, row = 3, comment = c2)
if (FALSE) {
saveWorkbook(wb, file = "createCommentExample.xlsx", overwrite = TRUE)
}