insertImage {openxlsx}	R Documentation

Insert an image into a worksheet

Description

Insert an image into a worksheet

Usage

insertImage(
  wb,
  sheet,
  file,
  width = 6,
  height = 3,
  startRow = 1,
  startCol = 1,
  units = "in",
  dpi = 300
)

Arguments

wb
A workbook object

sheet
A name or index of a worksheet

file
An image file. Valid file types are: jpeg, png, bmp

width
Width of figure.

height
Height of figure.

startRow
Row coordinate of upper left corner of the image

startCol
Column coordinate of upper left corner of the image

units
Units of width and height. Can be "in", "cm" or "px"

dpi
Image resolution used for conversion between units.

Author(s)

Alexander Walker

See Also

insertPlot()

Examples

## Create a new workbook
wb <- createWorkbook("Ayanami")

## Add some worksheets
addWorksheet(wb, "Sheet 1")
addWorksheet(wb, "Sheet 2")
addWorksheet(wb, "Sheet 3")

## Insert images
img <- system.file("extdata", "einstein.jpg", package = "openxlsx")
insertImage(wb, "Sheet 1", img, startRow = 5, startCol = 3, width = 6, height = 5)
insertImage(wb, 2, img, startRow = 2, startCol = 2)
insertImage(wb, 3, img, width = 15, height = 12, startRow = 3, startCol = "G", units = "cm")

## Save workbook
## Not run: 
saveWorkbook(wb, "insertImageExample.xlsx", overwrite = TRUE)

## End(Not run)

[Package openxlsx version 4.2.5 Index]