openxlsx {openxlsx}	R Documentation

xlsx reading, writing and editing.

Description

openxlsx simplifies the the process of writing and styling Excel xlsx files from R and removes the dependency on Java.

Details

The openxlsx package uses global options, most to simplify formatting. These are stored in the op.openxlsx object.

openxlsx.bandedCols

FALSE

openxlsx.bandedRows

TRUE

openxlsx.borderColour

"black"

openxlsx.borders

"none"

openxlsx.borderStyle

"thin"

openxlsx.compressionLevel

"9"

openxlsx.creator

""

openxlsx.dateFormat

"mm/dd/yyyy"

openxlsx.datetimeFormat

"yyyy-mm-dd hh:mm:ss"

openxlsx.headerStyle

NULL

openxlsx.keepNA

FALSE

openxlsx.na.string

NULL

openxlsx.numFmt

NULL

openxlsx.orientation

"portrait"

openxlsx.paperSize

9

openxlsx.tabColour

"TableStyleLight9"

openxlsx.tableStyle

"TableStyleLight9"

openxlsx.withFilter

NA Whether to write data with or without a filter. If NA will make filters with writeDataTable and will not for writeData

See the Formatting vignette for examples.

Additional options

See Also

vignette("Introduction", package = "openxlsx")

vignette("formatting", package = "openxlsx")

writeData()

writeDataTable()

write.xlsx()

read.xlsx()

op.openxlsx()

for examples

[Package openxlsx version 4.2.5 Index]