openxlsx {openxlsx}	R Documentation

xlsx reading, writing and editing.

Description

openxlsx simplifies the the process of writing and styling Excel xlsx files from R and removes the dependency on Java.

Details

The openxlsx package uses global options to simplify formatting:

options("openxlsx.borderColour" = "black")

options("openxlsx.borderStyle" = "thin")

options("openxlsx.dateFormat" = "mm/dd/yyyy")

options("openxlsx.datetimeFormat" = "yyyy-mm-dd hh:mm:ss")

options("openxlsx.numFmt" = NULL)

options("openxlsx.paperSize" = 9) ## A4

options("openxlsx.orientation" = "portrait") ## page orientation

See the Formatting vignette for examples.

Additional options

options("openxlsx.compressionLevel" = "9") ## set zip compression level, default is "1".

See Also

vignette("Introduction", package = "openxlsx")

vignette("formatting", package = "openxlsx")

writeData

writeDataTable

write.xlsx

read.xlsx

for examples

[Package openxlsx version 4.1.5 Index]