temp_xlsx {openxlsx}	R Documentation

helper function to create tempory directory for testing purpose

Description

helper function to create tempory directory for testing purpose

Usage

temp_xlsx(name = "temp_xlsx")

Arguments

name
for the temp file

[Package openxlsx version 4.2.5 Index]